+++ src/Recipe/Step/Render.php
@@ -25,6 +25,7 @@ declare(strict_types=1);
 
 namespace Teknoo\East\Website\Recipe\Step;
 
+use Psr\Http\Message\MessageInterface;
 use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\StreamFactoryInterface;
@@ -55,13 +56,16 @@ class Render
      * @param mixed $objectInstance
      */
     public function __invoke(
-        ServerRequestInterface $request,
+        MessageInterface $message,
         ClientInterface $client,
         string $template,
         ?string $objectViewKey = null,
         $objectInstance = null
     ): self {
-        $viewParameters = $request->getAttribute(ViewParameterInterface::REQUEST_PARAMETER_KEY, []);
+        $viewParameters = [];
+        if ($message instanceof ServerRequestInterface) {
+            $viewParameters = $message->getAttribute(ViewParameterInterface::REQUEST_PARAMETER_KEY, []);
+        }
 
         if (!empty($objectViewKey)) {
             $viewParameters = \array_merge($viewParameters, [$objectViewKey => $objectInstance]);